+++ pkg/service/service.go
@@ -661,10 +661,10 @@ func (s *Service) populateReverseNatMapV2FromV1(ipv4, ipv6 bool) error {
 		copyRevnatEntries := func(key bpf.MapKey, value bpf.MapValue) {
 			if v == v4 {
 				origKey := key.(*lbmap.RevNat4Key)
-				newKey = lbmap.NewRevNat4KeyV2(origKey.GetKey())
+				newKey = lbmap.NewRevNat4KeyV2(uint16(origKey.GetKey()))
 			} else {
 				origKey := key.(*lbmap.RevNat6Key)
-				newKey = lbmap.NewRevNat6KeyV2(origKey.GetKey())
+				newKey = lbmap.NewRevNat6KeyV2(uint16(origKey.GetKey()))
 			}
 
 			err := v2Map.Update(newKey, value)
